Relax config import to make it usable.
The os.path.isfile("config.py") check is too strict: it makes blitiri
usable from the command line only if it's ran in the directory where it
lives.

There's no point in doing that, let's just relax the import.
The chances of a config module collision are too low to be worth the
usability problem.


Signed-off-by: Alberto Bertogli <albertito@gmail.com>